+++ Window.py
@@ -13,8 +13,10 @@ class Window(Frame.Window):
         pygame.mixer.music.set_volume(main.config.config["Volume"] / 100)
         self.main = main
         self.key = 0
+        self.pressedMouse = False
 
     def update(self):
+        self.pressedMouse = False
         self.key = 0
         super().update()
         if self.keys[pygame.K_LALT] and self.keys[pygame.K_F4]:
@@ -24,22 +26,25 @@ class Window(Frame.Window):
         super().handleEvent(event)
         if event.type == pygame.KEYDOWN:
             self.key = event.key
-    
+        if event.type == pygame.MOUSEBUTTONDOWN:
+            self.pressedMouse = True
+
     def onOptionalSurfaceRender(self):
         self.screenSurfaces = []
     
     def exit(self):
-        finished = False
-        x0 = -250
-        x1 = 0
-        while not finished:
-            x0 += self.dt * 5
-            x1 += self.dt * 4
-            if x1 >= 1500:
-                finished = True
-            pygame.draw.polygon(self.surface, (0, 0, 0), [[x0, 0], [x1, 1080], [0, 1080], [0, 0]])
-            self.updateClock()
-            self.updateDisplay()
+        if self.main.config.config["ExitAnimation"]:
+            finished = False
+            x0 = -250
+            x1 = 0
+            while not finished:
+                x0 += self.dt * 5
+                x1 += self.dt * 4
+                if x1 >= 1500:
+                    finished = True
+                pygame.draw.polygon(self.surface, (0, 0, 0), [[x0, 0], [x1, 1080], [0, 1080], [0, 0]])
+                self.updateClock()
+                self.updateDisplay()
         pygame.quit()
         self.main.config.config["DebugScreenActive"] = self.main.debugScreenActive
         self.main.config.config["FPSLimit"] = self.main.window.fpsLimit